# specs/001-service-scheduling/contracts/openapi.yaml
openapi: 3.1.0
info: { title: Unified Service Scheduler, version: 1.0.0 }
paths:
  /api/v1/appointments:
    post:
      summary: Atomically create and confirm an appointment
      parameters:
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string, minLength: 8, maxLength: 128 } }
        - { in: header, name: X-Correlation-ID, required: false, schema: { type: string, maxLength: 128 } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateAppointmentRequest' } } }
      responses:
        '201':
          description: Appointment persisted and confirmed, or the original result replayed for the same idempotency key and request
          headers:
            Location: { schema: { type: string, format: uri-reference } }
            X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' }
          content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      summary: List appointments
      parameters:
        - { $ref: '#/components/parameters/CorrelationId' }
        - { in: query, name: dealershipId, schema: { type: string, format: uuid } }
        - { in: query, name: startDate, schema: { type: string, format: date-time } }
        - { in: query, name: endDate, schema: { type: string, format: date-time } }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/AppointmentStatus' } }
        - { in: query, name: page, schema: { type: integer, minimum: 0, default: 0 } }
        - { in: query, name: size, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { in: query, name: sort, schema: { type: string, default: 'startTime,asc' } }
      responses:
        '200':
          description: Page of matching appointments
          headers: { X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentPage' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/appointments/{appointmentId}:
    get:
      summary: Retrieve an appointment
      parameters:
        - { in: path, name: appointmentId, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/CorrelationId' }
      responses:
        '200':
          description: Appointment details
          headers: { X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/availability:
    get:
      summary: Advisory preview; appointment creation remains authoritative
      parameters:
        - { $ref: '#/components/parameters/CorrelationId' }
        - { in: query, name: dealershipId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: serviceTypeId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: startTime, required: true, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Current advisory availability
          headers: { X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                type: object
                required: [available, advisory]
                properties: { available: { type: boolean }, advisory: { type: boolean, const: true } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  parameters:
    CorrelationId:
      in: header
      name: X-Correlation-ID
      required: false
      schema: { type: string, maxLength: 128 }
  headers:
    CorrelationId:
      description: Validated incoming correlation identifier or a generated identifier
      schema: { type: string, maxLength: 128 }
  schemas:
    CreateAppointmentRequest:
      type: object
      additionalProperties: false
      required: [customerId, vehicleId, dealershipId, serviceTypeId, startTime]
      properties:
        customerId: { type: string, format: uuid }
        vehicleId: { type: string, format: uuid }
        dealershipId: { type: string, format: uuid }
        serviceTypeId: { type: string, format: uuid }
        startTime: { type: string, format: date-time, description: Offset required and normalized to UTC }
    AppointmentStatus: { type: string, enum: [CONFIRMED, CANCELLED] }
    ResourceReference:
      type: object
      required: [id, uri]
      properties:
        id: { type: string, format: uuid }
        uri: { type: string, format: uri-reference, description: Canonical identifier URI; dereferenceability is not guaranteed in Scenario A }
    Appointment:
      type: object
      required: [id, customer, vehicle, dealership, serviceType, technician, serviceBay, startTime, endTime, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        customer: { $ref: '#/components/schemas/ResourceReference' }
        vehicle: { $ref: '#/components/schemas/ResourceReference' }
        dealership: { $ref: '#/components/schemas/ResourceReference' }
        serviceType: { $ref: '#/components/schemas/ResourceReference' }
        technician: { $ref: '#/components/schemas/ResourceReference' }
        serviceBay: { $ref: '#/components/schemas/ResourceReference' }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/AppointmentStatus' }
        createdAt: { type: string, format: date-time }
    AppointmentPage:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content: { type: array, items: { $ref: '#/components/schemas/Appointment' } }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
    Problem:
      type: object
      required: [type, title, status, detail, instance, code, timestamp, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, description: Request path }
        code: { type: string, examples: [BOOKING_CONFLICT] }
        timestamp: { type: string, format: date-time }
        correlationId: { type: string }
  responses:
    BadRequest:
      description: Malformed, invalid, inconsistent request, or idempotency payload mismatch
      headers: { X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Referenced resource or appointment not found
      headers: { X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: No available combination or concurrent request won the race
      headers: { X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    InternalError:
      description: Unexpected operational failure without internal details
      headers: { X-Correlation-ID: { $ref: '#/components/headers/CorrelationId' } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }